requirements.txt
Django==1.11.3
django-appconf==1.0.2
django-imagefit==0.5
django-imagekit==4.0.1
django-versatileimagefield==1.7.1
olefile==0.44
pilkit==2.0
Pillow==4.0.0
pytz==2017.2
six==1.10.0
